add bitly user info retrieval function
This commit adds a function to retrieve Bitly user info using the
Bitly API.The function uses the user's API token and group GUID to
make a GET request to the API and return the user info as a JSON
object in text format. A separate function is also included to print
the user info to the console.