Add count_clicks() function
Retrieves the total number of clicks for a given Bitlink URL.
The function had one optional argument that is loaded from the virtual
environment with the help of dotenv library. Within the function a user
is asked to input a Bitlink URL for which total clicks should be
counted. The function returns the total number of clicks as an integer
or an error message if Bitlink service is unavailable.