Add is_bitlink() function
Add a new function is_bitlink that checks if a given URL is a Bitlink.
The function prompts the user to input a URL and then uses the
"urlparse" library to parse the URL and extract the first part of the
"/path". If the extracted part of the link is 'bit.ly' the function
returns "True", indicating that the URL is a Bitly link. Otherwise the
function returns "False".