Implement logic to chose an action with a given URL
Add a script that prompts a user to input a URL and then either
shortens it using shorten_link() function or count the clicks on the
URL if the URL is a Bitly link. The code checks whether the provided
link is a Bitly link using the is_bitlink() function.