Delete excess code (connection.py and URL check)
Delete file connection.py with function that retrieves Bitly user info,
delete try-except block from shorten_link function that
handles the case when a wrong URL was entered by a user. Removal is
necessary because the code as well blocks link shortening when the
given URL just is not working now. User wants to be able to shorten
inactive URLs too.
Also correct the README file in accordance with the changes.