Change is_bitlink() function
Now is_bitlink() function checks if the given URL is a Bitlink by
means of Bitly's "Retrieve a Bitlink" service. The function returns
True if the service response is OK, otherwise it returns False. This
approach allows to check correctly not only "bit.ly" links, but custom
domain links as well.