Correct some errors: is_bitlink() and prints
Delete response.raise_for_status() from the function is_bitlink, because
response.raise_for_status() was raising an exception,
while the function's result was supposed to return True or False
depending on whether response.ok or not. Thus, previously
the function worked correctly only if the status code was 200.
Descriptions of prints were confused, correct this.